% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entr.R
\name{entr}
\alias{entr}
\title{Entropy of the food environment}
\usage{
entr(df_res, df_tot = df_res)
}
\arguments{
\item{df_res}{the restaurant dataframe}

\item{df_tot}{the restaurant dataframe for the "scale", can be restaurant of all resturants, or the same as only e.g. take-home, by default is the same as df_res}
}
\description{
Entropy of the food environment
}
