% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entr_score.R
\name{entr_score}
\alias{entr_score}
\title{calculate the entropy score (by brainstorm)}
\usage{
entr_score(home_res, ready_res, df_tot)
}
\arguments{
\item{home_res}{the take-home restaurants environement}

\item{ready_res}{the ready to eat restaurants environment}

\item{df_tot}{the df_tot as in the entr function.}
}
\description{
calculate the entropy score (by brainstorm)
}
