% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthScore.R
\name{healthScore}
\alias{healthScore}
\title{Calculate healthiness Score}
\usage{
healthScore(df_res, df_tot = df_res, score, normalise = c("Ncati",
  "Nall"), method = c("simple", "expFun", "threshold"), threshold = 2)
}
\arguments{
\item{df_res}{the dataframe of the restaurant environment(count within a buffer)}

\item{df_tot}{dataframe of the restaurant environment for normalisation. By default same as the df_res}

\item{score}{the healthiniess score}

\item{normalise}{if normalise is Ncati, normalise by the number of restaurant categotris, e.g. if there are 3 fastfood and 5 cafe, the number of categories is 2. If the normalise is Nall, then normalise by the total number of resturant, in the previous example: 8.}

\item{method}{method to calculate healthiness score, can be from c("simple", "expFun", “threshold")}
}
\description{
Calculate healthiness Score
}
\details{
all the 0 restaurants records are NA.  if method is "simple", healthScore = sum(NumberRestau \* score). The method "expFun" calculate healthiness score with an exponential function of the number of restaurants sum(log(numberOfrestaurant+1) \* score). The "threshold" method treats more than a certain number (i.e. threshold) restaurants as the certain number of resturants
}
