% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMtable.R
\name{LMtable}
\alias{LMtable}
\title{generate table to compare healthscore and AHEI, MDS, DAS under linear regression}
\usage{
LMtable(healthscore1, xlabplot = "sum of healthiness score")
}
\arguments{
\item{healthscore1}{healthiness score}

\item{xlabplot}{the name of the xlab}
}
\description{
generate table to compare healthscore and AHEI, MDS, DAS under linear regression
}
