% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming.R
\name{renaming}
\alias{renaming}
\title{Replace names of a dataset
 @param dataset the dataset}
\usage{
renaming(dataset, names2replace = "loc", replacenames)
}
\arguments{
\item{replacenames}{the names to replace to, a list match the length of names to be replaced}

\item{name2replace}{greplstyle name to replace, e.g. "loc"}
}
\description{
Replace names of a dataset
 @param dataset the dataset
}
