% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1.R
\name{S1}
\alias{S1}
\title{healthiness score1}
\usage{
S1(df_res, score)
}
\arguments{
\item{df_res}{the dataframe of the resturant environment(count within a buffer)}

\item{score}{the healthiniess score}
}
\description{
healthiness score1
}
\details{
calculated as sum(score * foodenv_count)
}
