% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero2NA.R
\name{zero2NA}
\alias{zero2NA}
\title{Replace rowsum of zero to NA}
\usage{
zero2NA(df, baseon)
}
\arguments{
\item{df}{the dataframe which needs to be processed}

\item{baseon}{the dataframe which the calcualtion is based on. It could be the same as the df to be processed.}
}
\description{
Replace rowsum of zero to NA
}
\details{
if there is no resturant for an agent, replace the number of resturant (i,e, 0) to NA
}
